Remove TinyMCE plugin
If we decide to readd this, we should reapproach it from scratch. We can
look at this commit for stuff to salvage/reuse, but given the focus is on
the Gutenblock, I think dedicating time to this feature for limited value isn't
worth it.

If someone wants this, open an issue and we'll consider it.

Fixes #797
