+++ client/actions/index.ts
@@ -7,4 +7,3 @@ export * from './highlighting';
 export * from './init';
 export * from './jobs';
 export * from './router';
-export * from './tinymce';
